requirements.txt
ollama
tqdm
python-magic
pillow
pytesseract
openai-whisper
chromadb
langchain
langchain-community
networkx
fastapi
uvicorn
pyyaml
scikit-learn
numpy
pandas
rich
typer
torch
python-multipart
markdown
jinja2
